fix(decoder): First pass of cleanup
